Move the current sensors out of src/battery/ into src/shunt/
Shunt.h, Shunts.cpp, BMW-SBOX.h and BMW-SBOX.cpp are current sensing, not
battery drivers. The S-BOX's consumers are the Kostal RS485 inverter (which
special-cases ShuntType::BmwSbox for its current reading) and Chargebyte CCS;
BMW-I3-BATTERY.cpp has no reference to it. Keeping the shunt framework under
src/battery/ makes the directory mean two different things, and any tooling
or build filter that treats src/battery/ as "the battery drivers" quietly
takes the current sensors with it.

A new src/shunt/ mirrors src/charger/: the base class, the registry, and the
one CAN shunt device. CHADEMO-SHUNTS stays in src/battery/ - it is part of
the Chademo driver, not a standalone sensor.

Pure move plus include-path updates (ten references, one of them the
"./Shunt.h" spelling in CHARGEBYTE-CCS.cpp). No code changes.
